refactor: simplify REST API property handling
The REST API and its Query API part are determined by the following properties:

* `zeebe.broker.gateway.enable` - considered enabled by default, disable by setting it to `false`
* `camunda.rest.enabled` - considered enabled by default, disable by setting it to `false`
* `camunda.rest.query.enabled` - considered disabled by default, enable by setting it to `true`

Providing no configuration, the REST API is enabled, the REST Query API is disabled.
If either `zeebe.broker.gateway.enable` or `camunda.rest.enabled` are set to `false`, the REST API
and thus also the REST Query API are disabled.
If all three properties are set to `true`, the whole REST API, including the Query API, is enabled.
The `zeebe.broker.gateway.enable` and `camunda.rest.enabled` are considered `true` if not provided.