docs: mark client query commands as experimental
Marks the query command for process instance search as experimental in the Java client
because the related API is disabled by default. Thus, the method is not yet intended
for general use and can be subject to change.

Remove the experimental status from user task commands since they are now established.