test: fix flaky test shouldPushForMultipleJobsCreated
Rarely, when we arrive at `jobStream.getActivatedJobs().forEach()` we don't have all activated jobs in the stream yet.
But in the meantime, they get added while iterating over them, causing `ConcurrentModificationException` to be thrown.

To solve this we shall wait first for the stream to have the expected count, before looping on the stream items.