fix: resetting a reader does not invalidate returned record
When a reader was reset, it would mutate "internal" state by resetting
buffers to wrap over nothing.
It turns out that this state is not really internal because the
`LoggedEvent` returned by `next` or `peekNext` wraps over those internal
buffers. This can cause bugs where a successfully read entry is
surprisingly invalidated when the reader is closed.

It is much safer if entries returned by the reader are remain valid
even after a reset.

To do this, reset now creates new buffers and `LoggedEvent`s. This
makes the fields non-final and requires allocation on reset but makes
usage much safer.