fix: ensure active call activity on activate child
When activating a child process instance, we must ensure that the call
activity instance (represented by the parentElementInstanceKey) is still
active and not interrupted.

This fixes a NullPointerException that could otherwise occur during the
activation of the child process instance if the call activity instance
is already gone from the state due to Terminate End Event.