refactor: pull ensureTenantIdSet method up
Message correlation is the first endpoint that requires tenant verification. In the distant future other endpoints must
also use this mechanism. Therefore, we should make it available in a more generic place. I chose for the ApiServices as
every service class extends from it. This also contains the utility method getDocumentOrEmpty.