fix: skip Optimize in most builds
Building Optimize in Zeebe and common pipelines caused
most of the time failing builds due to timeouts. Likely
to longer buildtime, and other unknowns.

To get the team time to work and fix such issues, we
ignore for now the Optimize part. The build time needs to
be improved, and Optimize needs to migrate
to the common parent.