fix: clean up multi-tenancy validation
This cleans up the validation. Before it would just throw an exception that got caught and turned into a generic
ProblemDetail. Now it actually creates it's own ProblemDetail.
There is some ordering in these validations. First it will check if the provided tenantId is actually valid. Afterwards
it'll check if the provided is tenant is authorized to make the request, and finally we will validate the request itself
.