fix: teach rejection to be event type agnostic
The rejection of a migration due to subscribing to a catch event was
previously specific to message catch events only. However, it's now also
relied upon for timer catch events, and others in the future.

Rather than making the error message specific, we can make it slightly
less specific by removing the catch event type from the rejection
reason. This is fine, because the cause still contains the exact reason
for the failure.