test: allow concurrent migration/correlation
There appears to be no tests that verifies multi partition message
subscription migration and correlation. Which seems like a big
oversight.

When making the test below multi-partition, moving the message
subscription to partition 3 (key1 => key3), and not awaiting the
migrated events before writing the publish message command, it
highlights that the correlated event is written with the old bpmn
process id.