fix: adjust unsupported catch event type rejection
The rejection reason for attempting to migrate an intermediate catch
event of an unsupported event type was:

> Migrating active elements with intermediate catch event of these types
> is not possible yet.

However, this intermediate catch event is itself active. It's not part of
another active element like intermediate catch events attached to
event-based gateways.

So, while the assertions could make sense for intermediate catch events
attached to event-based gateways, those are not yet supported for
migration.

We can therefore greatly simplify the validation logic necessary:
- simply check the bpmn element type and bpmn event type of the
  currently active element instance