feat: store process key by process id and version tag + add find method
Adds a new column family to store process definition keys by process id and version tag.
The column family does not store the entire object, but just a foreign key to reduce the required storage space.
For the same reason, a new entry is only created if the version tag is not blank (it's an optional property).

Also, a new method is added to the process state that allows to find processes by process id and version tag.
This will be needed to find the correct version of a called process if the new "version tag" binding type is used.