fix: reject migrate element for eventbased gateway
We don't yet support migrating intermediate catch events that are
attached to event-based gateways. So, now that we support message
intermediate catch events, we must reject this case as unsupported
explicitly.

It's difficult to write a test case for this, as the activation of the
intermediate catch event is achieved through events that are written as
part of the completion of the event-based gateway, which in turn is the
result of correlating a message. The problem here is that we can only
use user commands (publish a message and migrate a process instance),
but cannot control the order of internal commands between these actions
easily.

As a workaround, we can use EngineRule.writeRecords to write some
events. Note that the events written with this method are not
automatically applied to the state, so we need to restart the engine to
achieve their effects.

By ensuring that the event-based gateway is completed (removed from the
state), and the intermediate catch event is activated when we process
the MIGRATE command, we can ensure that we encounter this specific case
that we wanted to reject.