refactor: align handling catch events
For intermediate catch events, there was some additional code added so
migrating (or subscribing/unsubscribing) of the event subscriptions
could be handled separately from the subscriptions for boundary events.

The need for this, was an line in the handleCatchEvents method where the
target element was expected to be an ExecutableActivity. However, in the
case of an intermediate catch event, it's an ExecutableCatchEvent which
is itself not an ExecutableActivity, i.e. events are not activities in
bpmn.

However, the code did not actually rely on the target element being an
activity. All it needs is to access the events that the element should
subscribe to. This is captured by a common interface between activities
and intermediate catch events: ExecutableCatchEventSupplier. This
interface provides exactly that access.

By using this, we can continue to rely on the `handleCatchEvents` method
for handling all catch events.