fix: ensure clock modifications are thread-safe
Removes the optimization to store `millis` so that we only have one
mutable field that requires synchronization.

Marks the `modification` field as volatile to ensure that modifications
are immediately visible to all threads.

This prevents confusing behavior, for example where we'd modify the
clock in the stream processor but other writing threads such as the
command api or scheduled tasks would not immediately observe this change
and thus write records with the wrong timestamp.