test: migrate active message event subprocesses
Migrating active event subprocesses is supported out of the box since we already handle message
catch event subscription migration with a generic implementation.

There is one thing to note though, when user wants to migrate an active message event subprocess
that contains same message name for the start event in both source and the target, they had to
provide a mapping. This is currently a limitation for all types of message subscriptions. Users
can still workaround this by providing a mapping for the related elements. In case of message event
subprocesses, users need to provide mapping for start events of event subprocesses. This limitation
needs to be documented.