build: unlearn spotless to ratchet from main
Ratcheting can be used to introduce spotless formatting gradually to a
project. However, our project already applies (and checks) the
formatting of all supported files (java, pom, markdown).

Locally, my machine is able to run `spotless:check` and `spotless:apply`
much faster (~4x faster) when ratchetFrom is disabled.

Possibly this is because of the size of our git repo. Note that spotless
itself is fast and uses incremental up to date checking and formatting
to avoid doing unnecessary work.