refactor: return immutable clock for observability
Exposes a new `clock` method on the stream processor to allow
introspection into the current stream clock. To avoid thread safety, it
returns an immutable clock which is frozen at the time of the call, but
also reports any applied modification.

To simplify the API, the `currentModification` method is moved into the
general `StreamClock` interface, and `UncontrolledStreamClock` returns
always `Modification.none()`.