test: pause processing before generating next key
This unflakes `MigrateProcessInstanceConcurrentNoBatchingTest` by
pausing the processing before generating a record key.

This test needs to generate a record key for use in some events that it
appends to the log as part of the set up of this test case.

Generating a next key with `DbKeyGenerator` is not a safe operation for
use in tests. It reads and changes a value in the state. The state is
also accessed concurrently by the stream processor. This can lead to
errors in committing the database transaction when both are trying to do
so simultaneously.

We can avoid this by pausing processing before generating the key.

An alternative approach would be to stop the engine first (which we do
after generating the key so we can write events that are applied during
replay), but this is not possible. If the engine is stopped, the state
is inaccessible.