feat: resolve called processes for binding type "versionTag"
If the binding type "versionTag" is used in a call activity, the particular version of the target process referenced by
the given version tag will be used.

Note that the same version tag can be assigned to multiple deployed versions of a process. In this case, the latest of
these versions will be used.

If no process is found for the given process ID and version tag, an incident will be raised.