refactor: record value and metadata length is always non-zero
This is a performance improvement to avoid calculating the length of
record value and metadata for every entry that's being appended.

Record values inherit from `ObjectValue` which always has non-zero
length due to the map header.
Record metadata is serialized via SBE which also requires a header and
thus always has non-zero length.