fix: set a limit to evaluationFailureMessage size
There was a bug where `evaluationFailureMessage` was failed to be indexed when it goes over
`keyword` data type's size. A limit is set to the related field to ignore the remaining part
of the keyword when it is above the size 1024. I chose number as 1024 because evaluation failure
message typically contains an error from a FEEL expression which is not expected to be too long.