feat: store decision key by id and version tag + add find method
Adds a new column family to store decision keys by decision id and version tag.
The column family does not store the entire object, but just a foreign key to reduce the required storage space.
For the same reason, a new entry is only created if the version tag is not blank (it's an optional property).

Also, a new method is added to the decision state that allows to find deployed decisions by decision id and version tag.
This will be needed to find the correct version of a called decision if the new "version tag" binding type is used in a
business rule task.