feat: resolve called decisions for binding type "versionTag"
If the binding type "versionTag" is used in a business rule task, the particular version of the target decision
referenced by the given version tag will be used.

Note that the same version tag can be assigned to multiple deployed versions of a decision. In this case, the latest of
these versions will be used.

If no decision is found for the given decision ID and version tag, an incident will be raised.