feat: always send user create command to deployment partition
At first I thought this was not neccessary. However, with the current state of Zeebe it's better to have one partition
be the single source of truth for user management. If we don't we could run into conflicts due to the distribution of
the commands.
For example, if 2 user create commands are sent shortly after another it could occur that they are both processed on a
different partition. If the processing happens before the distribution both partition will accept the command, create a
user and respond to the client that creation was successful. However, distribution will get stuck as both partitions
cannot create the user from the other partition, as a user with this specific username already exists in the state.