test: verify message event subprocess is not re-created after migration
With the current structure of handling catch events in `ProcessInstanceMigrationMigrateProcessor`,
when there is an unmapped catch event in the target, we subscribe to that event. However, it is a
wrong behaviour for interrupting event subprocesses. While migrating interrupting event
subprocesses, a new subscription for the start event shouldn't be created because interrupting
subprocess flow should be taken only once.