feat: store form key by form id and version tag + add find method
Adds a new column family to store form keys by form id and version tag.
The column family does not store the entire object, but just a foreign key to reduce the required storage space.
For the same reason, a new entry is only created if the version tag is not blank (it's an optional property).

Also, a new method is added to the form state that allows to find deployed forms by form id and version tag.
This will be needed to find the correct version of a linked form if the new "version tag" binding type is used in a
user task.