feat: resolve linked forms for binding type "versionTag"
If the binding type "versionTag" is used in a user task, the particular version of the target form referenced by the
given version tag will be used.

Note that the same version tag can be assigned to multiple deployed versions of a form. In this case, the latest of
these versions will be used.

If no form is found for the given form ID and version tag, an incident will be raised.