test: ensure mappings can target the same element
Even though we don't allow instance migration where a single boundary
event is the target of multiple mappings, we should allow several
mappings to target the same element in general.

There's a use case where an activity is removed from a model, and some
of the activities are "moved back" while others remain. Another use case
is that the model has changed so much that several different elements
need to be merged together.