fix: ensure boundary events are not merged
Rather than a simple duplication check, this constructs a Map so we can
easily report all mappings that reference the same target boundary
event.

Constructing the map is slightly more difficult, as we need to collect
both the target element id as well as all source element ids that
reference it. This makes it a candidate for Collectors.groupingBy.

I've introduced a small record Mapping, to use instead of something like
Tuple, as it clarifies that the grouping by is classified by the
targetElementId while we collect the sourceElementIds.