fix: use message subscription from state
One of the previously written tests was failing because it couldn't find
the MIGRATED event for the message subscription which was just CREATED.
The assertion worked by looking for the MIGRATED event based on the key
of the Message Subscription record. Somehow, the event had a different
key then the message subscription itself.

The problem is that the MIGRATE command is distributed from the process
instance's partition, which doesn't know the key of the message
subscription. As a workaround (and to provide a useful distribution
key), the process message subscription's key is used instead. This looks
slightly strange on the log, but that's sadly the way it has to be until
we fix the root cause. However, it caused this failing test.

When we process the MIGRATE command, we simply reuse this key for the
event that is written. Additionally, we just accept the data from the
command as the truth. However, in a distributed (partitioned) system, we
don't know whether the message subscription still exists on the
partition when the command to migrate it is received and being
processed. As the context may change in our system, command processors
must always consider the current state to decide how to process the
command.

For migrate message subscription commands, the processor must check
whether the subscription exists in the state, and then mark it as
migrated. It's also important to not blindly take all data from the
command. It may contain outdated information. Instead, we should use the
information from the state about the subscription for the MIGRATED
event, and adjust only the relevant data: bpmn process id +
interrupting.

Note that the subscription is not found by key but by element instance
key and message name.

If the subscription cannot be found at all, we should reject the command.
In that case, we should not acknowledge the distributed command to
ensure the command is retried. For example, when the MIGRATE command
arrives on this partition BEFORE the CREATE command had. (note currently
the order of distributed commands cannot be guaranteed although this may
change in the future).

This does raise an issue though. If the message is already correlated,
the migrate command would be retried indefinitely. This case would
always exist, even if we added the ordering guarantee, because we allow
correlating a message to a process instance for which the message
subscription is currently migrating.

So, we need to accept that the least worst solution is to always
acknowledge the command. This may mean we miss migrating a message
subscription if we hadn't distributed the CREATE command yet. We can
resolve this as a separate bug of course.