fix: copy record before appending
The record is read from the state, modified, and then appended. When
appending it using the state writer, the event is applied to the state.
The specific event applier (MessageSubscriptionMigratedApplier) reads
the subscription from the state again. This second read reuses the
object that was provided in the initial read from state. So the
modifications we made are undone.

Instead, we can copy the record before passing it to the state writer.