build: replace protolock with buf CLI
Adds a `buf` configuration file to the root of the project, where we
define the modules which will be built and compared for breaking
changes.

Right now, this is:
- `zeebe/dynamic-config`
- `zeebe/gateway-protocol`

The `zeebe/dynamic-config` module is configured to only check for WIRE
breaking changes (i.e. binary incompatibility), not source changes.

The `zeebe/gateway-protocol` module is configured to check for both
binary and source code breaking changes.

If you want to allow a breaking change, the PR must be labeled with the
`breaking/protobuf` label, indicating there is an allowed breaking
change to one of the protocols.

This means we do not run the check on merge group, but only on push and
pull requests. To react to label changes, the main CI workflow is
updated to add `labeled` and `unlabeled` triggers for `pull_request`
events.

done