fix: ensure catch events are not merged
We already checked that boundary events cannot be merged by mapping two
boundary events into a single one. This would mean the element instance
is subscribed twice to the same boundary event. A similar restriction
should be added for event-based gateways. It should not be possible to
merge catch events attached to an event-based gateway.

We can ensure this by making a small change to the method that checks
the boundary event mappings for referencing the same target boundary
event more than once. All we have to do is look at ALL THE EVENTS of the
element rather than only at the boundary events.

This means that the rule is actually more general than I initially
thought. But it makes sense, any element instance that is subscribed to
multiple catch events should not all of the sudden merge those events
into fewer catch events.

We can see that this is correct, if we go over all elements that can
subscribe to catch events:
- For event-based gateway, the only events it can be subscribed to are
  the attached intermediate catch events.
- For tasks, the only events it can be subscribed to are the attached
  boundary events
- For embedded subprocesses, the events can be either boundary events,
  or event-subprocesses
- For event-subprocesses, the events can be other event-subprocesses
  only
- For processes, the events can be event-subprocesses or regular start
  events (but the process won't be subscribed to start events)
- For intermediate catch events, the event is itself only
- For boundary events, the event is itself only
- For receive task, the event is itself in addition to boundary events

In all these cases, the rule holds: when migrating an element instance,
the catch events it is subscribed to should not be merged by multiple
mapping instructions targeting the same catch event more than once.

This now means the rule is already available as well for when we support
migrating event-subprocesses.