fix: align catch event detached rule
Just like the previous commit, the rules regarding the detachment of
boundary events from an activity and the detachment of intermediate
catch events from an event-based gateway can be generalized.

All catch events must stay attached to the element they are attached to
during migration.

While this means the error message is less specific, the rule is must
simpler to check and maintain, as the details are no longer specific to
the type of catch event.

While is mostly a refactor, it does change behavior as the error
messages are different. Additionally, this rule would be immediately
applicable to event-subprocesses as well once these are supported.