test: migrate business rule task
A process instance may be waiting on a Business Rule Task due to an Incident created as a result of
an invalid decision id. This can be resolved by the user, by deploying a new version of the
process and migrating the process instance over to it. When resolving the incident, the business
rule task will use the decision referenced in the target process definition and evaluate the
decision based on target process definition's properties.

Second, a Business Rule Task might be based on a job worker implementation. It is simply the same
case as the Service Task migration where we migrate the related job as well.