test: migrate script task
A process instance may be waiting on a Script Task due to an Incident created as a result of
an invalid feel expression. This can be resolved by the user, by deploying a new version of the
process and migrating the process instance over to it. When resolving the incident, the Script Task
will use the updated feel expression in the target process definition and evaluate it.

Second, a Script Task might be based on a job worker implementation. It is simply the same
case as the Service Task migration where we migrate the related job as well.