test: ensure eventsubprocess is not detached
Like other catch events, we must ensure that the event subprocess is not
detached from an element that is subscribed to it.

No rejection is provided for this case yet, but it will be automatically
taken care of with #21615.

To avoid having to remember writing a test for it there, I think it's
better to already add the test case here.