feat: validate user create requests
Adds a validation and mapping from the UserWithPasswordRequest DTO to an internal object (CreateUserRequest). When we
have a CreateUserRequest the request is validated and we can assume all fields are correct.

The validation is quite straightforward. It checks for all fields that they are not empty or blank. Email validation is
a bit more tricky. It's always hard to determine if an email is valid or not. Instead of coming up with something myself
 I used Apache's commons-validator. This has a built-in email validator. I did have to pull in a new dependency for
 this.