test: ensure eventsubprocess is not merged
Like other catch events, we must ensure that the event subprocess is not
merged for an element that is subscribed to them.

When an element is subscribed to two event subprocesses, we cannot merge
these together into a single event subprocess. So we have to reject the
migration.

No rejection is provided for this case yet, but it will be automatically
taken care of with #21615.

To avoid having to remember writing a test for it there, I think it's
better to already add the test case here.