test: await for incident migrated
The test case should wait for the incident to be migrated before
resolving the incident.

While not strictly necessary, by waiting for the incident to be migrated
we're clear about the expectations before continuing with the next step.

It's not strictly necessary because the `.migrate()` call already awaits
the successful migration (the success expectation), and the incident is
migrated directly as part of the instance migration. But, it's good to
make these assumptions clear in the test case.