fix: make ALREADY_EXISTS return a 409, instead of a 400
The definition of a 409 Conflict is:
This response is sent when a request conflicts with the current state of the server.

When you try to create and it already exists that is a conflict with the current state of the server. Therefore, this
code is more accurate than the generic 400 Bad Request.