fix: make SBE code generation thread safe
The SBE tool used for code generation is not thread safe in itself, as
it uses some static/shared properties. This is not an issue with
single-threaded Maven builds, but when using a multi-threaded Maven
build, with multiple modules, this can cause some builds to fail during
code generation.

To avoid this, we change the use of `exec:java` to `exec:exec`, which
will call `java` as a separate process. It would be nicer if `exec:java`
would allow forking, but it unfortunately does not.

At the same time, this commit centralizes a lot of the SBE settings and
makes it easier to add SBE code generation to a sub-module.

One thing to note, the execution to generate SBE is initially not bound
to any phase, as the plugin itself is used in other modules. So when you
want to add code generation to your module, you have to configure that
specific `generate-sbe` execution of the plugin and bind it to a phase.

(cherry picked from commit af4368ea8bc312ad42aa1999834cd2b95da6c012)