fix: skip initialization of partition ring until topology is initialized
This fixes a bug where the partition ring was incorrectly initialized from the topology.
Checking for `null` is not enough, we must also account for the situation where cluster topology is available but not
yet initialized with cluster size and partition count.

Without this fix, we'd initialize with partition count 0 and then never update it again, eventually causing
`ArithmeticException: / by zero` when trying to look up a partition that wasn't known before.