feat: benchmark worker send messages
In our realistic benchmarks we need to complete receive tasks
and message catch events, with message with special correlation keys
based on PI payload. Before these catch events normal a send task exists
to symbollically send a message to a different system and after that
we want to receive as message again.

We can mimik this with a worker completing the send task and also publish a message
for the follow up receive task.

As the correlation key must be unique to the process instance otherwise
it will not delivered, we need to get the correlation key out of the PI payload
which we get in the send worker.