refactor: use signal subscription from the state while unsubscribing
A method introduced to CatchEventBehavior to be able to filter signal events by signal subscription
instance retrieved from the state. It is needed to keep track of signal subscriptions instances to
migrate. While visiting signal subscriptions to unsubscribe, we will keep the ones that should be
migrated in a list. Afterward, we will migrate those instances.

It is the same refactoring as done for timer instance migrations. If we don't do this refactoring,
we need to handle retrieving signal subscription records to migrate from the state by their element
id.