feat: implement signal subscription migration
Adding Signal to allowed types:
Signal catch event is added to supported boundary event types.

Unsubscribing:
We go over signal subscriptions in the state to unsubscribe from them. During the iteration, we
check 2 things:
1) If the signal subscription found in the state has target process definition key. That means, we
already subscribed to this signal above in the `handleCatchEvents()` method because that signal
exists in the target process. We shouldn't unsubscribe from it.
2) If the boundary event of the signal subscription is mapped. If that is the case, we shouldn't
unsubscribe from it, and we store the  signal subscription in a list to migrate it.

Migrating:
Lastly, we go over the list of signal subscriptions to migrate. We update fields that needs to be
changed during the migration. They are process definition key, catch event it and bpmn process id.
Please note that we do not update signal name. That feature will be supported later.