test: verify re-subscribing to the same signal name
In the added test, the signal name is the same both in the source and the target. Currently, it
fails. Because in the state, we use `signal name` and `subscription key` (instance key of the
element that the signal event belongs to) as the column key. That overrides the existing signal
subscription in the state and leads to incorrect behaviour. It will be fixed in the following
commits.