fix: allow re-subscribing to the same signal name
Unsubscribing from signal events first in method `handleCatchEvents()` resolved the issue. Because
when we remove the signal subscription from the state, we can later create a new one while
subscribing to it. If we subscribe first, we override the signal subscription existing in the state.

At this point, `handleCatchEvents()` method is not easily readable. Because it has many
responsibilities and now, the order of execution changed due to unsubscribing from signal events
first. In the following commits, I will refactor the method to make it more readable.