refactor: rely on distribution key order only
This removes the insertion key that was added under the assumption that we can't rely on distribution key ordering.
We'd prefer to eventually generate distribution keys internally to ensure uniqueness and ordering.
Until then, we can accept the risk that queue order is undefined in the (currently not existing) case that someone is
using old distribution keys in a queue.