refactor: handle message publish response
We need to make sure that the message is published successfully
If it is not delived and we complete our corresponding job, we have a
stuck process instance.

To avoid this we track the message publish response and retry the job in
failure cases.

Cleans up a bit the handling and makes the completion delay calculation
a bit more clever to consider also the message publish response time