feat: add partition scoped meter registry
This solves three problems:
1. How to remove meters when partitions are removed
2. How to reset meters when partitions change roles
3. How to set the partition id label for every meter

We do this by creating a new registry on every partition transition.
This allows us to define common labels (i.e. partition id) and to remove all meters specific to that partition and role.

The partition registry is wrapped over the general broker registry such that all meters are forwarded to the correct
backend automatically.