test: verify migrating error boundary events
When migrating the process instance the error boundary event will be migrated to the target process
regardless of mapping instruction is provided for the error boundary element or not. After the
instance is migrated, the error code will be updated as in the target process since it is static
value.

Following test cases are added:
- Migrating error boundary event when a mapping instruction is provided for boundary events
- Migrating error boundary event when no mapping instruction is provided for boundary events
- Migrating multiple error boundary events