feat: new camunda performance dashboard
Contribute new Camunda performance dashboard this dashboard should allow
to give a holistically overview of the general performance of the Camunda
platform.

Currently this involves, Zeebe, Operate, and Elasticsearch mainly. Further
needs to be added.

It gives a general overview of the performance and healthiness of running components
in the first section.

In later sections panels show latency and throughput related metrics.