fix: don't correlate to start event if catch event for same processId already correlated
When a message start event and a message intermediate event have the same message name, publishing a message correlated
to both of these. When these both exists within the same process definition (same processId), we only want to correlate
to the intermediate catch event. By passing this Subscriptions object around we ensure that this is the case, as the
correlateToMessageStartEvents methods checks this object for correlated subscriptions with the same processId.