feat: define new requests for changing cluster configuration
There are two ways to scale
1. By providing the new clusterSize
2. By providing the set of members to add or remove.

Both requests allow to change all of or a subset of cluster size, partition
count and replication factor.

In addition, a new request for force remove is provided where we can provide
the set of brokers to remove, instead of the members to retain.