fix: use initial deployment key in metadata of duplicate resources
When a deployed resource has not changed and is marked as duplicate, use the deployment key of the deployment the
resource was initially deployed with (instead of the deployment key of the current deployment) in the resource's
metadata that is added to the deployment record.
This is done to be consistent with the deployment key that was persisted initially in the actual resource record.