feat: create AuthorizationAddPermissionProcessor
This processor processes the ADD_PERMISSION command. It has to check we know the owner and if we do, write the
PERMISSION_ADDED event.

The command must be distributed. For this distribution we make use of the IDENTITY distribution queue. All identity
related commands will get distributed to this queue. If we don't do this we will run into issues where distributions
could overtake each other which results in state inconsistencies across partitions.