feat: store permissions in state
This commit introduces a new method on the DbAuthorizationState. This method will create or add permissions to an
Authorization. It will look for an existing Authorization for this owner + resource + permission. If it didn't find one
it will create one. If it did find one, it will update it with the new permissions.