refactor: extract treePath cache
Create a new class for the treePath cache,
FlowNodeInstanceTreePathCache.

The FlowNodeInstanceTreePathCache comprises the corresponding
multi-level HashMap (with SoftHashMap), and the resolvment via a given
treePathResolver.

A new record was created to encapsulate the necessary values like,
partitionId, recordKey, flowScopeKey and processInstanceKey. This
allows to reduce the dependency to the Zeebe protocol. Allows us to move
the cache to the common module.

As the resolver is available via DI the cache can be easily tested via
unit-tests.

Added new tests and documentation to cover and describe the behavior of
the cache.