refactor: switch pending distribution CF usages to retriable distribution CF
The pending distribution will be replaced by the retriable CF. Therefore, all usages of the pending CF need to switched
to the retriable CF.

This commit likely switches too much. Some of it will probably be switched back later. However, this gives me a clean
base to work upon in the next commits.