test: ensure catch event type cannot be changed
It's currently possible to migrate a catch event subscription to a catch
event of a different event type. This is unexpected as the underlying
subscription isn't really changing type. For example, a message boundary
event that is migrated to a timer boundary event will still be triggered
by a correlated message and no timer will be created for it.

Therefore, we need to reject this case. It can happen in:
- boundary events
- event-based gateway
- event-subprocess