fix: reject on missing process in timer trigger
Rejects the Timer.TRIGGER command when the related
process definition cannot be found.

Otherwise, the command can run into an exception that
leads to banning the related process instance. For start
timer events, this means banning the last process instance
that was created by the start timer trigger.

This adds another safeguard for very rare use cases that shouldn't occur regularly.

No test case is added for this since the scenario cannot be easily created in any way.
The test case needs the following to verify the new rejection occurs:

* A timer instance in the timer state that links to a process definition that doesn't exist.
* A Timer.TRIGGER command for that timer instance on the log.

Usually, all timer instances are removed when the related process definitions are deleted.
Thus, by regular means available in our test setups, it is not easily possible to create this.