refactor: verify multi-instance flow scope cannot be changed
While verifying multi-instance flow scope. I realized that the current logic produces wrong error
message. With the previous precondition, for element instance MULTI_INSTANCE_BODY, it first checks
if the expected flow scope id (`targetProcessId`) matches the actual ('sub'). It is not. Then, it
moves to the next if statement and assigns inner activity's flow scope ('serviceTask2') to the
expected flow scope. Eventually it rejects the command but with a wrong error message. Error
message says that the actual flow scope should be 'serviceTask2' which is wrong.

When we change the order of if statements, the rejection message is corrected.