feat: reject migration that changes element multi instance status
While testing the case where a non-multi-instance activity is marked as multi-instance during the
migration, it allowed the migration with the current precondition implementation. It is because the
precondition doesn't know if the element instance is part of a multi instance or not. Thus, it sees
the service task inside the multi instance as the same as the service task in the source and allows
the migration. That is a wrong behaviour. I updated the precondition and used multi instance loop
counter to detect whether currently migrated element instance is part of a multi instance or not.